% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneSNP}
\alias{geneSNP}
\title{Gene SNP for multiple loci with mutiple alleles}
\format{A matrix with columns:
\describe{
\item{pheRes}{The phenotypic trait response, with three categorical levels:
C, D and M.}
\item{snp10001}{The SNP data for locus snp10001}
\item{snp10003}{The SNP data for locus snp10003}
}}
\source{
Toy sample data
}
\usage{
geneSNP
}
\description{
Toy sample data
}
\examples{
\dontrun{
 geneSNP
}
}
\keyword{datasets}
