% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneSNP2}
\alias{geneSNP2}
\title{Gene SNP for multiple loci with mutiple alleles}
\format{A data frame with columns:
\describe{
\item{phenRes}{The phenotypic trait response, with three categorical levels:
C, D and M.}
\item{snp10001}{The SNP data for locus snp10001}
\item{snp10002}{The SNP data for locus snp10002}
\item{snp10003}{The SNP data for locus snp10003}
\item{snp10005}{The SNP data for locus snp10005}
\item{snp10008}{The SNP data for locus snp10008}
}}
\source{
Toy sample data
}
\usage{
geneSNP2
}
\description{
Toy sample data 2
}
\examples{
data(geneSNP2)
\dontrun{


}
}
\keyword{datasets}
