% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gws.R
\name{gws}
\alias{gws}
\title{Genowide computation and visualization}
\usage{
gws(hap, rep)
}
\arguments{
\item{hap}{The data from genowide phenotype records and SNP, in data frame format
First column is the phenotype, followed by all genotype SNPs.
The data scoure: from "SNPassoc" package, but phenotype data are simulated from binary to 3 levels}

\item{rep}{The number of replication of cross-validation in calculating gscore
(In genowide computation, larger than 10 rep will be slow)}
}
\value{
The generated plot, showing gscores of all loci, the actual gscores and most
strongly related loci can be get from summary.gws function
}
\description{
Computer gscore for every locus and use ggplot to plot all the scores, data need preparation
}
\examples{
:
data(hapData)
gws(hapData,5)


}
