% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hapData}
\alias{hapData}
\title{The data from genowide phenotype records and SNP, in data frame format}
\format{A data frame with columns:
\describe{
\item{V1}{The SNP data for sample V1}
\item{V2}{The SNP data for sample V2}
\item{V3}{The SNP data for sample V3}
\item{V4}{The SNP data for sample V4}
\item{V5}{The SNP data for sample V5}
\item{V6}{The SNP data for sample V6}
\item{V7}{The SNP data for sample V7}
\item{V8}{The SNP data for sample V8}
\item{V9}{The SNP data for sample V9}
\item{V10}{The SNP data for sample V10}
\item{V11}{The SNP data for sample V11}
\item{V12}{The SNP data for sample V12}
\item{V13}{The SNP data for sample V13}
\item{V14}{The SNP data for sample V14}
\item{V15}{The SNP data for sample V15}
\item{V16}{The SNP data for sample V16}
\item{V17}{The SNP data for sample V17}
\item{V18}{The SNP data for sample V18}
\item{V19}{The SNP data for sample V19}
\item{V20}{The SNP data for sample V20}
\item{V21}{The SNP data for sample V21}
\item{V22}{The SNP data for sample V22}
\item{V23}{The SNP data for sample V23}
\item{V24}{The SNP data for sample V24}
\item{V25}{The SNP data for sample V25}
\item{V26}{The SNP data for sample V26}
\item{V27}{The SNP data for sample V27}
\item{V28}{The SNP data for sample V28}
\item{V29}{The SNP data for sample V29}
\item{V30}{The SNP data for sample V30}

}}
\source{
Package SNPassoc
}
\usage{
hapData
}
\description{
The data from genowide phenotype records and SNP, in data frame format
}
\examples{
data(hapData)
\dontrun{
}
}
\keyword{datasets}
