% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.SNP.R
\name{is.SNP}
\alias{is.SNP}
\title{Decide whether the data is of type "SNP"}
\usage{
is.SNP(SNP)
}
\arguments{
\item{SNP}{The result data returned from prep() function, an object of class factor and SNP,
it contains levels of genotype and attribute of "alleleNames"}
}
\value{
TURE if input is of class 'SNP', FALSE otherwise
}
\description{
Decide whether the data is of type "SNP"
}
\examples{

data(geneSNP)
s3 <- prep(geneSNP$snp10003)
is.SNP(s3)
TRUE


}
