% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gws.R
\name{plot.gws}
\alias{plot.gws}
\title{Plot visualization of genome-wide computation
Plot function for gScore type data}
\usage{
plot.gws(gScore)
}
\arguments{
\item{gScore}{The result data returned from genome wide computation of function gws, in data frame format
First column is gSCore, followed by the related names of the loci in second column.}
}
\value{
The generated plot using ggplot2, showing gScores of all loci
}
\description{
Plot visualization of genome-wide computation
Plot function for gScore type data
}
\examples{

data(hapData)
gScore <- gws(hapData,5)
plot.gws(gScore)

}
