% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpgrou.R
\name{snpgrou}
\alias{snpgrou}
\title{Calculate metrics for how one or multiple locus is(are) related to phenotype}
\usage{
snpgrou(x, rep)
}
\arguments{
\item{x}{A data frame containing one or multiple SNP data for every loucs, the first column
is the phenotype data, follow by SNP data column. If SNP data not alraedy in "SNP" type, this
function use prep function to prepare the data.}

\item{rep}{Integer, the number of replication for the cross-validation procedures.}
}
\value{
The results of gScore for each locus
}
\description{
This function use linear discriminant analysis to classify response phenotype according to
}
\examples{
snpgrou(geneSNP, 100)
#  snp10001  snp10002  snp10003  snp10005  snp10008
#  0.2946199 0.8362573 0.3413450 0.3141520 0.3190643


}
