% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gws.R
\name{summary.gws}
\alias{summary.gws}
\title{Summarize the actual gScore and other results from genome wide computation of gws()
The input gScore must obtain from gws() return}
\usage{
\method{summary}{gws}(gScore, threshold = 0.6)
}
\arguments{
\item{gScore}{The result data returned from genome wide computation of function gws, in data frame format
First column is gSCore, followed by the related names of the loci in second column.}

\item{threshold}{Define high gSore ss values that greater than this parameter}
}
\description{
Summarize the actual gScore and other results from genome wide computation of gws()
The input gScore must obtain from gws() return
}
\examples{

data(hapData)
gScore <- gws(hapData, 5)
summary(gScore)
[1] "Found gScore values greater than 0.60 and responsible loci in the data:  "
[1] "High gScore 0.787037 at chrosomal position: rs4661627"
[1] "High gScore 0.675926 at chrosomal position: rs12749939"
[1] "Loci below are excluded from the calculation because snnp data only contain one allel"
[1] "rs6691424"  "rs7553868"  "rs10922638" "rs12097782" "rs6428582"  "rs12739867"
[7] "rs189401"   "rs282036"   "rs526519"   "rs2818971"  "rs12756456"
summary(gScore, 0.7)
[1] "Found gScore values greater than 0.70 and responsible loci in the data:  "
[1] "High gScore 0.787037 at chrosomal position: rs4661627"
[1] "Loci below are excluded from the calculation because snnp data only contain one allel"
[1] "rs6691424"  "rs7553868"  "rs10922638" "rs12097782" "rs6428582"  "rs12739867"
[7] "rs189401"   "rs282036"   "rs526519"   "rs2818971"  "rs12756456"
}
