% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryGws.R
\name{summaryGws}
\alias{summaryGws}
\title{Summarize the actual gScore and other results from genome wide computation of gws()
The input gScore must obtain from gws() return}
\usage{
summaryGws(gScore, threshold = 0.6)
}
\arguments{
\item{gScore}{The result data returned from genome wide computation of function gws, in data frame format
First column is gSCore, followed by the related names of the loci in second column.}

\item{threshold}{Define high gSore ss values that greater than this parameter}
}
\value{
charcter Loci below are excluded from the calculation

#'@references
González JR1, Armengol L, Solé X, Guinó E, Mercader JM, Estivill X, Moreno V(2007).
SNPassoc: an R package to perform whole genome association studies. Bioinformatics. 2007
Mar 1;23(5):644-5. Epub 2007 Jan 31. DOI: 10.1093/bioinformatics/btm025.
}
\description{
Summarize the actual gScore and other results from genome wide computation of gws()
The input gScore must obtain from gws() return
}
\examples{

data(hapData)
gScore <- gws(hapData, 5)
summaryGws(gScore)
#Found gScore values greater than 0.60 and responsible loci in the data:
#High gScore 0.787037 at chrosomal position: rs4661627"
#High gScore 0.675926 at chrosomal position: rs12749939"
#Loci below are excluded from the calculation because snnp data only contain one allel
#rs6691424 rs7553868  rs10922638 rs12097782 rs6428582
#rs12739867" rs189401   rs282036   rs526519   "rs2818971
#rs12756456
# summaryGws(gScore,0.7)
#Found gScore values greater than 0.70 and responsible loci in the data:
#High gScore 0.787037 at chrosomal position: rs4661627
#Loci below are excluded from the calculation because snnp data only contain one allel
#rs6691424  rs7553868  rs10922638 rs12097782 rs6428582
#rs12739867 rs189401   rs282036   rs526519   rs2818971
# rs12756456
}
