% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySNP.R
\name{summarySNP}
\alias{summarySNP}
\title{Summarize the frequency and percentage of genotypes and alleles in the SNP data
The input SNP data need prepared by prep() function}
\usage{
summarySNP(x)
}
\arguments{
\item{x}{The result data returned from prep function. it is a object of type "SNP" and "factor"
it contains levels of genotype and attribute of "alleleNames"}
}
\value{
A data frame contains the allele frequency and percentage
}
\description{
Summarize the frequency and percentage of genotypes and alleles in the SNP data
The input SNP data need prepared by prep() function
}
\examples{
data(geneSNP)
s3 <- prep(geneSNP$snp10003)
summarySNP(s3)
#This data contains Genotypes:
#Genotyepes frequency percentage
#1        A/T        15      30.61
#2        C/C         1       2.04
#3        C/T         1       2.04
#4        G/C        16      32.65
#5        G/G        16      32.65
#
#
#This data contains Alleles:
#Alleles frequency petcentage
#1       A        15      18.52
#2       G        32      39.51
#3       C        18      22.22
#4       T        16      19.75
}
\references{
González JR1, Armengol L, Solé X, Guinó E, Mercader JM, Estivill X, Moreno V(2007).
SNPassoc: an R package to perform whole genome association studies. Bioinformatics. 2007
Mar 1;23(5):644-5. Epub 2007 Jan 31. DOI: 10.1093/bioinformatics/btm025.
}
