% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupView.R
\name{groupView}
\alias{groupView}
\title{Visualization of association of multiple loci with phenotype}
\usage{
groupView(genesnp)
}
\arguments{
\item{genesnp}{The data frame of mutiple loci SNP data.
First column is the phenotype, followed by all genotype SNPs.}
}
\value{
The generated plot, showing how well the SNPs predicate the phenotype, if there are
strongly related loci in the data set then the plot will show well separated grouos, but
which locus is most strongly related is not showing
}
\description{
Use lda function to train linear discriminant model from SNP data, then use this model to
predicate phenotype, use ggplot visualize how well the predication separate the phenotype
to groups
}
\examples{
require(ggplot2)
data(geneSNP)


}
