% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gws.R
\name{gws}
\alias{gws}
\title{Genome-wide computation}
\usage{
gws(hap, rep)
}
\arguments{
\item{hap}{The data from genome wide phenotype records and SNP, in data frame format
First row is the phenotype, followed by all genotype SNPs in all loci.
The data scoure: from "SNPassoc" package, but phenotype data are simulated from binary to 3 levels}

\item{rep}{The number of replication of cross-validation in calculating gScore
(In genome-wide computation, larger than 10 rep will be slow)}
}
\value{
The object of type "gScore", with attribute of "filt", the actual gScores and other
results can be get from summary.gws function
}
\description{
Compute gScore for every locus, data need preparation
}
\examples{

data(hapData)
gScore <- gws(hapData,3)


}
