% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hapData}
\alias{hapData}
\title{The data from genowide phenotype records and SNP, in data frame format}
\format{A data frame with columns:
\describe{
\item{group}{The phenotypic trait response, with three categorical levels:
1, 1 and 3.}
\item{rs10399749}{The SNP data for locus rs10399749}
\item{rs11260616}{The SNP data for locus rs11260616}

}}
\source{
Package SNPassoc
}
\usage{
hapData
}
\description{
The data from genowide phenotype records and SNP, in data frame format
}
\examples{
data(hapData)
\dontrun{
}
}
\keyword{datasets}
