% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\title{Prepare the data for the analysis purpose}
\usage{
prep(x, delim = NULL)
}
\arguments{
\item{x, }{numeric data of all obeservation of nucleotides in single locus}

\item{delim}{character that separate the original nucleotides data of two homologue in one locus, could be empty}
}
\value{
object belong to "snp" and "factor", with attributes of alleleNames
}
\description{
Transform the data to class "snp" object, nucleotides separated with "/"
}
\examples{


}
