% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\title{Prepare the data for the analysis purpose}
\usage{
prep(x, delim = NULL)
}
\arguments{
\item{x, }{Factors, data of all obeservation of alleles in single locus}

\item{delim}{character that separate the alleles in original data in one locus, could be empty}
}
\value{
object belong to "SNP" and "factor", with attributes of alleleNames
}
\description{
Transform the data to class "SNP" object, with alleles separated with "/"
}
\examples{
  s1<-prep(geneSNP$snp10001,delim="-") # This column original delimiter "-", replaced by "/"
  s2<-prep(geneSNP$snp10002,delim=",") # This column original delimiter "," replaced by "/"
  s3<-prep(geneSNP$snp10003)           # This column original data no delimiter, add "/"


}
